chore: move oauth components to folder
